Remove structure.lib.php and display_structure.inc.php
