+++ templates/structure/table_header.phtml
@@ -14,7 +14,13 @@ if ($GLOBALS['cfg']['NumFavoriteTables'] > 0) {
     <thead>
         <tr>
             <th class="print_ignore"></th>
-            <th><?php echo PMA_sortableTableHeader(__('Table'), 'table'); ?></th>
+            <th><?php echo PMA\Template::get('structure/sortable_header')->render(
+                   array(
+                       'title' => __('Table'),
+                       'sort' => 'table',
+                       'initial_sort_order' => 'ASC'
+                   )
+                ); ?></th>
             <?php if ($replication): ?>
                 <th><?php echo __('Replication'); ?></th>
             <?php endif; ?>
@@ -23,40 +29,94 @@ if ($GLOBALS['cfg']['NumFavoriteTables'] > 0) {
             </th>
             <!-- larger values are more interesting so default sort order is DESC-->
             <th>
-                <?php echo PMA_sortableTableHeader(__('Rows'), 'records', 'DESC'); ?>
+                <?php echo PMA\Template::get('structure/sortable_header')->render(
+                    array(
+                        'title' => __('Rows'),
+                        'sort' => 'records',
+                        'initial_sort_order' => 'DESC'
+                    )
+                ); ?>
                 <?php echo PMA_Util::showHint(PMA_sanitize(
                     __('May be approximate. Click on the number to get the exact'
                         . ' count. See [doc@faq3-11]FAQ 3.11[/doc].'))); ?>
             </th>
             <?php if (!($GLOBALS['cfg']['PropertiesNumColumns'] > 1)): ?>
-                <th><?php echo PMA_sortableTableHeader(__('Type'), 'type'); $cnt++; ?></th>
-                <th><?php echo PMA_sortableTableHeader(__('Collation'), 'collation'); $cnt++; ?></th>
+                <th><?php echo PMA\Template::get('structure/sortable_header')->render(
+                        array(
+                            'title' => __('Type'),
+                            'sort' => 'type',
+                            'initial_sort_order' => 'ASC'
+                        )
+                    ); $cnt++; ?></th>
+                <th><?php echo PMA\Template::get('structure/sortable_header')->render(
+                        array(
+                            'title' => __('Collation'),
+                            'sort' => 'collation',
+                            'initial_sort_order' => 'ASC'
+                        )
+                    ); $cnt++; ?></th>
             <?php endif; ?>
 
             <?php if ($GLOBALS['is_show_stats']): ?>
                 <!-- larger values are more interesting so default sort order is DESC -->
-                <th><?php echo PMA_sortableTableHeader(__('Size'), 'size', 'DESC'); $cnt++; ?></th>
+                <th><?php echo PMA\Template::get('structure/sortable_header')->render(
+                        array(
+                            'title' => __('Size'),
+                            'sort' => 'size',
+                            'initial_sort_order' => 'DESC'
+                        )
+                    ); $cnt++; ?></th>
                 <!-- larger values are more interesting so default sort order is DESC -->
-                <th><?php echo PMA_sortableTableHeader(__('Overhead'), 'overhead', 'DESC'); $cnt++; ?></th>
+                <th><?php echo PMA\Template::get('structure/sortable_header')->render(
+                        array(
+                            'title' => __('Overhead'),
+                            'sort' => 'overhead',
+                            'initial_sort_order' => 'DESC'
+                        )
+                    ); $cnt++; ?></th>
             <?php endif; ?>
 
             <?php if ($GLOBALS['cfg']['ShowDbStructureComment']): ?>
-                <th><?php echo PMA_sortableTableHeader(__('Comment'), 'comment'); $cnt++; ?></th>
+                <th><?php echo PMA\Template::get('structure/sortable_header')->render(
+                        array(
+                            'title' => __('Comment'),
+                            'sort' => 'comment',
+                            'initial_sort_order' => 'ASC'
+                        )
+                    ); $cnt++; ?></th>
             <?php endif; ?>
 
             <?php if ($GLOBALS['cfg']['ShowDbStructureCreation']): ?>
                 <!-- newer values are more interesting so default sort order is DESC -->
-                <th><?php echo PMA_sortableTableHeader(__('Creation'), 'creation', 'DESC'); $cnt++; ?></th>
+                <th><?php echo PMA\Template::get('structure/sortable_header')->render(
+                        array(
+                            'title' => __('Creation'),
+                            'sort' => 'creation',
+                            'initial_sort_order' => 'DESC'
+                        )
+                    ); $cnt++; ?></th>
             <?php endif; ?>
 
             <?php if ($GLOBALS['cfg']['ShowDbStructureLastUpdate']): ?>
                 <!-- newer values are more interesting so default sort order is DESC -->
-                <th><?php echo PMA_sortableTableHeader(__('Last update'), 'last_update', 'DESC'); $cnt++; ?></th>
+                <th><?php echo PMA\Template::get('structure/sortable_header')->render(
+                        array(
+                            'title' => __('Last update'),
+                            'sort' => 'last_update',
+                            'initial_sort_order' => 'DESC'
+                        )
+                    ); $cnt++; ?></th>
             <?php endif; ?>
 
             <?php if ($GLOBALS['cfg']['ShowDbStructureLastCheck']): ?>
                 <!-- newer values are more interesting so default sort order is DESC -->
-                <th><?php echo PMA_sortableTableHeader(__('Last check'), 'last_check', 'DESC'); $cnt++; ?></th>
+                <th><?php echo PMA\Template::get('structure/sortable_header')->render(
+                        array(
+                            'title' => __('Last check'),
+                            'sort' => 'last_check',
+                            'initial_sort_order' => 'DESC'
+                        )
+                    ); $cnt++; ?></th>
             <?php endif; ?>
         </tr>
     </thead>
